% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofoo.r
\name{foofoo}
\alias{foofoo}
\title{Test function for help}
\usage{
foofoo()
}
\description{
The purpose of this function is to test out \code{help} and \code{?} from
devtools.
}
\examples{
stopifnot(foofoo() == 'You called foofoo.')
}
