% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofoo.r
\name{testExampleMacro}
\alias{testExampleMacro}
\title{Function with a custom macro in its examples}
\usage{
testExampleMacro()
}
\description{
Test that examples can use custom Rd macros.
}
\examples{
a <- \exampleval{303}
}
