% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofoo.r
\docType{package}
\name{testHelp-package}
\alias{testHelp}
\alias{testHelp-package}
\title{testHelp: some title
Some description}
\description{
testHelp: some title
Some description
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
